Move hydrate break message dictionary to separate folder
